Modifique el mainpage falta cambiar nombres y un poco el formato
