+++ src/App.css
@@ -6,8 +6,10 @@ body {
 .app {
     display: flex;
     height: 100vh;
-    max-width: 1300vh;
-    margin-left: auto;
-    margin-right: auto;
+    width: 200vh;
+    max-height: 100vh;
+    /* max-width: 1300vh; */
+    /* margin-left: auto; */
+    /* margin-right: auto; */
     padding: 0 10px;
 }
