+++ src/components/Leftside/Leftbar.css
@@ -1,17 +1,21 @@
 .leftbar {
-    border-right: 1px solid var(--twitter-background);
-    flex: 0.3;
+    width: 100%;
     min-width: 250px;
-    margin-top: 20px;
+    max-width: 93%;
+    margin-right: 30vh;
     padding-left: 20px;
     padding-right: 20px;
+    display: flex;
+    justify-content: space-between;
+    position: static;
 }
 
-.leftbar__twitterIcon {
-    color: var(--twitter-color);
-    font-size: 30px !important;
+.leftbar__twitterIcon > svg {
+    filter: invert(62%) sepia(33%) saturate(7190%) hue-rotate(179deg) brightness(95%) contrast(97%);
+    height: 30px;
     margin-left: 20px;
-    margin-bottom: 20px;
+    margin-top: 15px;
+    /* margin-bottom: 10px; */
 }
 .leftbar__twitterIcon:hover {
     cursor: pointer;
