+++ src/components/Leftside/LeftbarOption.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import "./LeftbarButtons.css";
+
+function LeftbarButtons({ text }) {
+  return (
+    <div className="leftbarButtons" >
+        <h2>
+            {text}
+        </h2>
+    </div>
+  );
+}
+
+export default LeftbarButtons;
