+++ src/components/Main.css
@@ -4,11 +4,12 @@ body {
 }
 
 .main {
-    display: flex;
-    height: 100vh;
+    display: block;
+    height: 95vh;
     align-content: center;
-    max-width: 1200px;
-    margin-left: auto;
-    margin-right: auto;
+    max-width: 99%;
+    max-height: 99%;
+    margin-left: 0;
+    margin-right: 0;
     padding: 0 10px;
 }
