+++ src/components/Middle/Middlebar.css
@@ -1,0 +1,33 @@
+.middlebar {
+    display: flex;
+    width: auto;
+    margin-top: 2%;
+}
+
+.middlebar__posts::-webkit-scrollbar {
+    display: none;
+}
+
+.middlebar__posts {
+    display: flex;
+    /* flex-flow: column wrap; */
+    overflow-x: scroll;
+    margin-left: 1%;
+    align-content: center;
+    -ms-overflow-style: none;
+    scrollbar-width: 0;
+}
+
+.middlebar__header {
+    top: 0;
+    background-color: white;
+    z-index: 100;
+    padding: 15px 20px;
+}
+
+.middlebar__header > h2 {
+    margin-bottom: 30%;
+    font-size: 30px;
+    font-weight: 800;
+    margin-left: 35%;
+}
