+++ src/components/Middle/Middlebar.js
@@ -5,7 +5,9 @@ import Post from './Post';
 import Elonmusk from "./images/elonmusk.jpeg";
 import Whoknows from "./images/whoknows.jpg";
 import Trump from "./images/Trump.jpg";
-import TrumpMeme from "./images/Trumpmeme.jpg"
+import TrumpMeme from "./images/Trumpmeme.jpg";
+import Toretomeme from "./images/Toretomeme.jpeg";
+import Will from "./images/Will.jpg"
 
 function Middlebar() {
 
@@ -25,6 +27,14 @@ function Middlebar() {
       text:"Bring down the wall!",
       avatar:Trump,
       image:TrumpMeme
+    },
+    {
+      displayName:"WillSmith",
+      username:"SlapPrince",
+      verified:false,
+      text:"Fastest hand of the west",
+      avatar:Will,
+      image:Toretomeme
     }
   ]
 
@@ -37,9 +47,11 @@ function Middlebar() {
     <div className='middlebar'>
         <div className='middlebar__header'>
             <h2>Home</h2>
-        </div>
         <Tweetbox />
+        </div>
+        <div className='middlebar__posts'>
         {renderOptions(postInfo)}
+        </div>
     </div>
   );
 }
