--- src/components/Middlebar.css
@@ -1,27 +1,0 @@
-.middlebar {
-    flex: 0.4;
-    width: auto;
-    vertical-align: center;
-    border-right: 1px solid var(--twitter-background);
-    min-width: fit-content;
-    overflow-y: scroll;
-    -ms-overflow-style: none;
-    scrollbar-width: none;
-}
-
-.middlebar::-webkit-scrollbar {
-    display: none;
-}
-
-.middlebar__header {
-    position: sticky;
-    top: 0;
-    background-color: white;
-    z-index: 100;
-    padding: 15px 20px;
-}
-
-.middlebar__header > h2 {
-    font-size: 20px;
-    font-weight: 800;
-}
