--- src/components/News.css
@@ -1,33 +1,0 @@
-.news {
-    margin-top: 20px;
-}
-
-.news:hover {
-    cursor: pointer;
-    background-color: #d1d1d1;
-    border-radius: 20px;
-}
-
-.news__header {
-    margin-left: 10px;
-
-    font-weight: 600;
-    font-size: 12px;
-    color: gray;
-}
-
-.news__content {
-    margin-left: 10px;
-    font-size: 15px;
-    font-weight: 700;
-    display:flex;
-    padding: 2px;
-    margin-bottom: 10px;
-}
-
-.news__image {
-    height: 60px;
-    margin-bottom: 10px;
-    margin-left: 20%;
-    margin-right: 5px;
-}
