+++ src/components/Rightside/Rightbar.css
@@ -1,36 +1,53 @@
 .rightbar {
-    flex: 0.3;
     width: auto;
+    display: flex;
+    max-height: 20vh;
+    max-width: 100%;
 }
 
 .rightbar__input {
     display: flex;
-    align-items: center;
+    align-content: center;
     background-color: rgb(223, 223, 223);
     padding: 10px;
     border-radius: 20px;
-    margin-top: 10px;
+    width: 300px;
+    margin-top: 25px;
     margin-left: 20px;
+    max-height: 50px;
 }
 
 .rightbar__input > input {
     border: none;
+    margin-left: 10px;
     background-color: rgb(223, 223, 223);
 }
 
 .rightbar__searchIcon {
     color: black;
+    margin-top: 12px;
+    margin-left: 15px;
 }
 
 .rightbar__widgetContainer {
-    margin-top: 15px;
+    display: flex;
+    margin-top: 10px;
     margin-left: 20px;
     padding: 20px;
     background-color: rgb(223, 223, 223);
     border-radius: 20px;
+    overflow-x: scroll;
+    -ms-overflow-style: none;
+    scrollbar-width: 0;
+    max-height: 80px;
 }   
 
+.rightbar__widgetContainer::-webkit-scrollbar {
+    display: none;
+}
+
 .rightbar__widgetContainer > h2 {
+    margin-right: 10px;
     font-size: 18px;
     font-weight: 800;
 }
